pyproject.toml
[tool.poetry]
name = "ecpay"
version = "0.1.0"
description = ""
authors = ["Marcus <s09203647@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "3.11"
Flask = "3.0.3"
requests = "2.31.0"
flask-wtf = "1.2.1"
cassandra-driver = "3.29.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
